Skip builtin.jq if env var JQ_SKIP_BUILTINS is set
Name resolution in libraries is basically O(n^2) in number of bindings
and the impact of builtin.jq on jq's startup time turns out to be
absolutely awful.

The situation is actually much better for inline definitions alongside
the 'program', which are at least discarded immediately if unreferenced.

Given this, it could be useful for jq programs that need would otherwise
waste a lot of time in linking because they are run frequently on small
(up to a few MB...) inputs to skip builtin.jq, which by definition can
be defined entirely in jq, albeit possibly using internal interfaces.

See also: https://github.com/stedolan/jq/issues/1380